personalization 수정

+++ src/main/java/com/jjikmeok/app/domain/personalization/dto/ActivityRecommendationProjection.java
@@ -16,5 +16,5 @@ public interface ActivityRecommendationProjection {
 
     Integer getActivityFavoriteCount();
 
-    Integer getRecommendScore();
+    Long getRecommendScore();
 }
